front end widgets

+++ src/LondonFencing/media/Apps/views/index.php
@@ -76,7 +76,12 @@ if (!empty($_POST)){
                         'w'        => 100,
                         'adaptive' => true
                     ),
-                    'large' => array(
+                    'home' => array(
+                        'l'        => 310, 
+                        'w'        => 960,
+                        'adaptive' => true
+                    ),
+                    'banner' => array(
                         'l'        => 800, 
                         'w'        => 800,
                         'adaptive' => false
@@ -116,7 +121,7 @@ if (!empty($_POST)){
                         }
                     }
                     else{
-                        $message = "Your photos could not be uploaded. Please check the size and file type";
+                        $message = "<h3>Your photos could not be uploaded. Please check the size and file type</h3>";
                     }
                 }
             break;
@@ -131,10 +136,27 @@ if (!empty($_POST)){
                 $db->free_result();
             }
             else{
-                $message = "You can not submit an empty tag. <em>Text required</em>";
+                $message = "<h3>You can not submit an empty tag. <em>Text required</em></h3>";
             }
             break;
         
+       case "deleteTag":
+            if (is_numeric($_GET['id']) && (int)$_GET['id'] > 0){
+                ($m->delete_tag((int)$_GET['id']) == 1) ? header('location:/admin/apps/media/index') : $message = "Your tag could not be deleted";
+            }
+            else{
+                $message = "<h3>You must select a tag to delete it</h3>";
+            }
+           break;
+       
+        case "delete":
+            if (is_numeric($_GET['id']) && (int)$_GET['id'] > 0){
+                ($m->delete_media_item((int)$_GET['id']) == 1) ? header('location:/admin/apps/media/index') : $message = "Your photo could not be deleted";  
+            }
+            else{
+                $message = "<h3>You must select an image to delete it</h3>";
+            }
+            break;
         default:
             
             break;
@@ -149,8 +171,8 @@ if (!empty($_POST)){
 include $root. "/admin/templates/header.php";
 ?>
 
-<h1>Media Manager</h1>
-<p>This allows the ability to post photos and other media on the public front end, and place collections of them on specific pages of your website.</p>
+<h1>Media Manager<?php if (isset($_GET['tag'])){ echo ' : Edit Tag';}?></h1>
+<p>This allows the ability to post photos on the public front end, and place collections of them on specific pages of your website.</p>
 <?php
 
 $page = (isset($_GET['page']) ? $_GET['page'] : 1);
@@ -160,80 +182,91 @@ $media_items = $m->get_media_list(false, $page, $tags);
 if (isset($message) && $message != ""){
     echo $message;
 }
+
 ?>
 <div class="boxStyle">
     <div class="boxStyleContent">
-        <div class="boxStyleHeading">
-            <h2>Upload Photos <span class="h2Details">(max: 10 x 2MB, types: .jpg, .png)</span></h2>
-            <div class="boxStyleHeadingRight">
-            <form action="/admin/apps/media/index<?php echo (isset($tags[0])? "?tag=".$_GET['tag'] : ''); ?>" method="post" name="frmUploadMedia" id="frmUploadMedia" enctype="multipart/form-data">
-            <input type="file" name="RQvalFILEUpload[]" id="RQvalFILEUpload"  multiple size="250" class="btnStyle blue" style="padding-top:2px"/>
-            <input type="submit" class="btnStyle blue" name="submitUpload" id="submitUpload" value="Upload" />
-            <input type="hidden" name="nonce" value="<?php echo Quipp()->config('security.nonce');?>" />
-            <input type="hidden" name="OPvalALPHTag_Name" value="<?php echo (isset($tags[0])? $tagRa[$tags[0]] : ''); ?>" />
-            <input type="hidden" name="dbaction" value="img" />
-            </form>
-            </div>
-        </div>
+    <div class="boxStyleHeading">
+            <h2><?php echo (!empty($tags) ? 'Tag: '.$tagRa[$tags[0]] : 'All Media'); ?></h2>
+    </div>
         <div class="clearfix">&nbsp;</div>
+        <div id="template">
 <?php
-        if (empty($tags)){
-  ?>          
-        <div class="boxStyleHeading">
-            <h2>Tags <span class="h2Details">Click Tag to Filter</span></h2>
-            <div id="upload_tags" class="boxStyleHeadingRight">
-            <form onsubmit="return false;">
-            <input type="text" class="mediaTags" id="upload_tags_input" class="upload_tags_input" name="newTags"/>
-            <div id="tagButtons" style="display:inline-block"><input type="button" class="btnStyle blue" id="submitTags" name="submitTags" value="Add Tags" /></div>
-            <input type="hidden" name="nonce" value="<?php echo Quipp()->config('security.nonce');?>" />
-            <input type="hidden" name="dbaction" value="tag" />
-            </form>
-            </div>
-            <div class="clearfix">&nbsp;</div>
-                    <ul id="tag_cloud">
-                <?php 
-                    array_map(function($tag_info) {
-                    $tag    = filter_var($tag_info['tag'], FILTER_SANITIZE_SPECIAL_CHARS, FILTER_FLAG_ENCODE_HIGH);
 
-                    $point  = 5;
-                    $weight = floor($tag_info['tag_count'] / $point) * $point;
-
-                    $active = (isset($_GET['tag']) && $_GET['tag'] == $tag_info['tagID'] ? ' active' : '');
-
-                    echo <<<EOL
-                <li class="tag_weight-{$weight}{$active}"><a href="/admin/apps/media/index?tag={$tag_info['tagID']}" title="{$tag_info['tag_count']} photos tagged">{$tag}</a></li>
-EOL;
-                }, $m->get_tag_cloud()); ?>
-            </ul>
-        </div>
-<?php
-        }
-        else{
+if (!empty($tags)){
 ?>
-        <div class="boxStyleHeading">
-            <h2><?php echo $tagRa[$tags[0]]; ?></h2>
-            <div class="boxStyleHeadingRight">
-            <form action="/admin/apps/media/index?tag=<?php echo $_GET['tag']; ?>" method="post" name="frmEditTag" id="frmEditTag" enctype="multipart/form-data">
-                <label for="RQvalALPHEdit_Tag"> Edit Tag </label>
-            <input type="text" id="RQvalALPHEdit_Tag" class="uniform" name="RQvalALPHEdit_Tag" value="<?php echo $tagRa[$tags[0]]; ?>" />
-           <div id="tagButtons" style="display:inline-block;vertical-align:middle"> <input type="submit" class="btnStyle blue" id="updateTag" name="updateTag" value="Update" /></div>
+        
+            <form action="/admin/apps/media/index?tag=<?php echo $_GET['tag']; ?>" method="post" name="frmEditTag" id="frmEditTag" class="tableEditorForm"enctype="multipart/form-data">
+                <table>
+                    <tr>
+                     <td style="width:30%"><label for="RQvalALPHEdit_Tag"> Edit Tag </label></td>
+                     <td style="width:60%"><input type="text" id="RQvalALPHEdit_Tag" class="uniform" name="RQvalALPHEdit_Tag" value="<?php echo $tagRa[$tags[0]]; ?>" /></td>
+                    <td><input type="submit" class="btnStyle blue" id="updateTag" name="updateTag" value="Update" /></td>
+                    </tr>
+                </table>
+                
             <input type="hidden" name="nonce" value="<?php echo Quipp()->config('security.nonce');?>" />
             <input type="hidden" name="OPvalNUMBTag_ID" value="<?php echo (isset($tags[0])? $_GET['tag'] : 0); ?>" />
             <input type="hidden" name="dbaction" value="uTag" />
             </form>
-            </div>
-            </div>
-            <div class="clearfix">&nbsp;</div>
 <?php        
         }
 ?>
-        <div class="clearfix">&nbsp;</div>
-        <div class="boxStyleHeading">
-            <h2>Photos <span class="h2Details">Displaying <span id="currently"> <?php echo $media_items->count(); ?></span> of <span id="total"><?php echo $media_items->getTotal(); ?></span></h2>
-        </div>
- <?php           
-            print "<div id=\"contactSheet\">";
-            print $m->getContactSheetMarkup($media_items);
+            <form action="/admin/apps/media/index<?php echo (isset($tags[0])? "?tag=".$_GET['tag'] : ''); ?>" method="post" name="frmUploadMedia" id="frmUploadMedia" class="tableEditorForm" enctype="multipart/form-data">
+            <table>
+                    <tr>
+                     <td style="width:30%"><label for="RQvalFILEUpload">Upload Photos</label><p>(max 10 x 2MB in a single upload. Use .png or .jpg. Home images should be optimized for 960 x 310)</p></td>
+                     <td style="width:60%"><input type="file" name="RQvalFILEUpload[]" id="RQvalFILEUpload"  multiple size="250" class="btnStyle" style="padding-top:2px"/></td>
+                    <td><input type="submit" class="btnStyle" name="submitUpload" id="submitUpload" value="Upload" /></td>
+                    </tr>  
+            </table>
+            <input type="hidden" name="nonce" value="<?php echo Quipp()->config('security.nonce');?>" />
+            <input type="hidden" name="OPvalALPHTag_Name" value="<?php echo (isset($tags[0])? $tagRa[$tags[0]] : ''); ?>" />
+            <input type="hidden" name="dbaction" value="img" />
+            </form>
+<?php
+        if (empty($tags)){
+  ?>          
+            <form onsubmit="return false;" class="tableEditorForm" id="frmNewTags">
+                <table>
+                    <tr>
+                     <td style="width:30%"><label for="upload_tags_input">Create New Tags</label></td>
+                     <td style="width:60%"><input type="text" class="mediaTags" id="upload_tags_input" class="upload_tags_input" name="newTags"/></td>
+                    <td><input type="button" class="btnStyle blue" id="submitTags" name="submitTags" value="Add Tags" /></td>
+                    </tr>
+                </table>
+            <input type="hidden" name="nonce" value="<?php echo Quipp()->config('security.nonce');?>" />
+            <input type="hidden" name="dbaction" value="tag" />
+            </form>
+            <div class="tableEditorForm" id="dvTagCloud">
+                <table>
+                    <tr>
+                     <td style="width:30%"><label>Current Tags</label><p>Select a Tag to filter</p></td>
+                     <td>
+                         <ul id="tag_cloud">
+                <?php 
+                    array_map(function($tag_info) {
+                    $tag    = filter_var($tag_info['tag'], FILTER_SANITIZE_SPECIAL_CHARS, FILTER_FLAG_ENCODE_HIGH);
+
+                    $active = (isset($_GET['tag']) && $_GET['tag'] == $tag_info['tagID'] ? ' active' : '');
+
+                    echo <<<EOL
+                <li><a href="/admin/apps/media/index?tag={$tag_info['tagID']}" title="{$tag_info['tag_count']} photos tagged">{$tag}</a></li>
+EOL;
+                }, $m->get_tag_cloud()); ?>
+            </ul>
+                     </td>
+                    </tr>
+                </table>
+                    
+            </div>
+            <div class="clearfix">&nbsp;</div>
+<?php
+        }
+?>        
+            <p style="padding: 5px 0 15px 0"><span class="contactSheetP">Photos</span> Displaying <span id="currently"> <?php echo $media_items->count(); ?></span> of <span id="total"><?php echo $media_items->getTotal(); ?></p>
+              <?php print "<div id=\"contactSheet\">";
+            print $m->getContactSheetMarkup($media_items, $tags);
             print "</div>";
 
             $url_parts  = parse_url(filter_input(INPUT_SERVER, 'REQUEST_URI', FILTER_UNSAFE_RAW));
@@ -247,12 +280,14 @@ EOL;
                 }
             }
 ?>
-        <div class="clearfix">&nbsp;</div>
+            <div class="clearfix">&nbsp;</div>
 <?php 
          if (!empty($tags)){
-            echo '<div class="boxStyleHeading">';
+            echo '<div class="boxReturn">';
             echo '<input type="button" class="btnStyle green" id="cancel" name="cancel" value="Return"  onclick="window.location=\'/admin/apps/media/index\'"/>';
+            echo '<input type="button" class="btnStyle red" id="deleteTag" name="deleteTag" value="Delete Tag"  onclick="javascript:confirmDelete(\'?action=deleteTag&id='.$_GET['tag'].'\')"/>';
              echo '</div>';
+             echo '<div class="clearfix">&nbsp;</div>';
          }
 ?>
 </div>
